requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
slowapi
python-jose
